pyproject.toml
[project]
name = "artifact"
version = "0.1.0"
description = ""
authors = [
    {name = "Wearemassive78",email = "massimiliano.pupillo@studio.unibo.it"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=3.0.0,<4.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "pyserial (>=3.5,<4.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "mypy (>=1.19.1,<2.0.0)",
    "pylint (>=4.0.4,<5.0.0)",
    "coverage (>=7.13.4,<8.0.0)",
    "types-pyyaml (>=6.0.12.20250915,<7.0.0.0)"
]
